ucsim-0.6-pre54: fix segfault problem in "info hw simif" and "timer set" commands; fix format specifiers in prints
git-svn-id: http://svn.code.sourceforge.net/p/sdcc/code/trunk/sdcc@11157 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ sim/ucsim/sim.src/uc.cc
@@ -2291,7 +2291,7 @@ cl_uc::get_rtime(void)
 unsigned long
 cl_uc::clocks_of_time(double t)
 {
-  return t * xtal;
+  return (unsigned long)(t * xtal);
 }
 
 int
@@ -2599,7 +2599,7 @@ cl_error_unknown_code::print(class cl_commander_base *c)
       /*cmd_fprintf(f,*/c->dd_printf(")");
     }
   else
-    /*cmd_fprintf(f,*/c->dd_printf("0x%06x", PC);
+    /*cmd_fprintf(f,*/c->dd_printf("0x%06x", AU(PC));
   /*cmd_fprintf(f,*/c->dd_printf("\n");
 }
